requirements.txt
numpy==1.25.1
Pillow==10.0.0
tqdm==4.65.0
